+++ src/mako/http/routing/middleware/SecurityHeaders.php
@@ -10,13 +10,6 @@ namespace mako\http\routing\middleware;
 use Closure;
 use mako\http\Request;
 use mako\http\Response;
-use mako\syringe\Container;
-use mako\view\ViewFactory;
-
-use function base64_encode;
-use function implode;
-use function json_encode;
-use function random_bytes;
 
 /**
  * Security headers middleware.
@@ -26,23 +19,9 @@ use function random_bytes;
 class SecurityHeaders implements MiddlewareInterface
 {
 	/**
-	 * Container.
-	 *
-	 * @var \mako\syringe\Container;
-	 */
-	protected $container;
-
-	/**
-	 * Report to.
-	 *
-	 * @var array|null
-	 */
-	protected $reportTo;
-
-	/**
 	 * Security headers.
 	 *
-	 * @var array|null
+	 * @var array
 	 */
 	protected $headers =
 	[
@@ -52,193 +31,17 @@ class SecurityHeaders implements MiddlewareInterface
 	];
 
 	/**
-	 * Should we only report content security policy violations?
-	 *
-	 * @var bool
-	 */
-	protected $cspReportOnly = false;
-
-	/**
-	 * Content security policy directives.
-	 *
-	 * @var array|null
-	 */
-	protected $cspDirectives =
-	[
-		'base-uri'    => ['self'],
-		'default-src' => ['self'],
-		'object-src'  => ['none'],
-	];
-
-	/**
-	 * Content security policy nonce.
-	 *
-	 * @var string|null
-	 */
-	protected $cspNonce;
-
-	/**
-	 * Content security policy nonce view variable name.
-	 *
-	 * @var string
-	 */
-	protected $cspNonceVariableName = '_csp_nonce_';
-
-	/**
-	 * Constructor.
-	 *
-	 * @param \mako\syringe\Container $container Container
-	 */
-	public function __construct(Container $container)
-	{
-		$this->container = $container;
-	}
-
-	/**
-	 * Builds the "Report-To" header value.
-	 *
-	 * @return string
-	 */
-	protected function buildReportToValue(): string
-	{
-		$endpoints = [];
-
-		foreach($this->reportTo as $endpoint)
-		{
-			$endpoints[] = json_encode($endpoint);
-		}
-
-		return implode(', ', $endpoints);
-	}
-
-	/**
-	 * Generates a random content security policy nonce.
-	 *
-	 * @return string
-	 */
-	protected function generateCspNonce(): string
-	{
-		return base64_encode(random_bytes(16));
-	}
-
-	/**
-	 * Returns the content security policy nonce.
-	 *
-	 * @return string
-	 */
-	protected function getCspNonce(): string
-	{
-		if($this->cspNonce === null)
-		{
-			$this->cspNonce = $this->generateCspNonce();
-		}
-
-		return $this->cspNonce;
-	}
-
-	/**
-	 * Builds the "Content-Security-Policy" header value.
-	 *
-	 * @return string
-	 */
-	protected function buildCspValue(): string
-	{
-		$directives = [];
-
-		foreach($this->cspDirectives as $name => $directive)
-		{
-			if($directive === true)
-			{
-				$directives[] = $name;
-
-				continue;
-			}
-
-			$directiveString = $name;
-
-			foreach($directive as $value)
-			{
-				switch($value)
-				{
-					case 'self':
-					case 'unsafe-inline':
-					case 'unsafe-eval':
-					case 'none':
-						$value = "'{$value}'";
-						break;
-					case 'nonce':
-						$value = "'nonce-{$this->getCspNonce()}'";
-						break;
-				}
-
-				$directiveString .= " {$value}";
-			}
-
-			$directives[] = $directiveString;
-		}
-
-		return implode('; ', $directives);
-	}
-
-	/**
-	 * Assigns a global view variable containing the content security policy nonce.
-	 */
-	protected function assignCspNonceViewVariable(): void
-	{
-		if($this->container->has(ViewFactory::class))
-		{
-			$this->container->get(ViewFactory::class)->assign($this->cspNonceVariableName, $this->getCspNonce());
-		}
-	}
-
-	/**
-	 * Should the CSP header be added to response?
-	 *
-	 * @param  \mako\http\Response $response Response
-	 * @return bool
-	 */
-	protected function shouldAddCspHeader(Response $response): bool
-	{
-		return $response->getType() === 'text/html';
-	}
-
-	/**
 	 * {@inheritdoc}
 	 */
 	public function execute(Request $request, Response $response, Closure $next): Response
 	{
 		$headers = $response->getHeaders();
 
-		if($this->reportTo !== null)
+		foreach($this->headers as $name => $value)
 		{
-			$headers->add('Report-To', $this->buildReportToValue());
+			$headers->add($name, $value);
 		}
 
-		if($this->headers !== null)
-		{
-			foreach($this->headers as $name => $value)
-			{
-				$headers->add($name, $value);
-			}
-		}
-
-		if($this->cspDirectives !== null)
-		{
-			$cspHeader = $this->buildCspValue();
-
-			if($this->cspNonce !== null)
-			{
-				$this->assignCspNonceViewVariable();
-			}
-		}
-
-		$response = $next($request, $response);
-
-		if($this->cspDirectives !== null && $this->shouldAddCspHeader($response))
-		{
-			$headers->add($this->cspReportOnly ? 'Content-Security-Policy-Report-Only' : 'Content-Security-Policy', $cspHeader);
-		}
-
-		return $response;
+		return $next($request, $response);
 	}
 }