+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Added `mime_type` validation rule.
 * Added `max_filename_length` validation rule.
 * Added `SecurityHeaders` middleware.
+* Added `ContentSecurityPolicy` middleware.
 * Added `FileSystem::getDiskSize()` method.
 * Added `FileSystem::getFreeSpaceOnDisk()` method.
 * Added `JSON::getStatus()` method.